poc_translator/requirements.txt
python>=3.9
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
scikit-learn>=1.1.0
xgboost>=1.6.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
torch>=1.12.0
pytorch-lightning>=1.8.0
tensorboard>=2.10.0
pyyaml>=6.0
umap-learn>=0.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
joblib>=1.1.0
tqdm>=4.64.0
# Optional: wandb for experiment tracking
# wandb>=0.13.0
